Conflicts:
	simcity.cc
	simcity.h
	simversion.h

+++ utils/float32e8_t.h
@@ -55,7 +55,7 @@ protected:
 			}
 			else
 			{
-				return _ild[x];
+				return _ild[(uint8)x];
 			}
 		}
 	}
